+++ relay/polkadot/src/ah_migration/phase1.rs
@@ -18,7 +18,6 @@
 
 use crate::*;
 use frame_support::traits::Contains;
-use pallet_rc_migrator::types::PortableFreezeReason;
 
 /// Contains all calls that are enabled during the migration.
 pub struct CallsEnabledDuringMigration;
@@ -139,38 +138,6 @@ pub fn call_allowed_status(call: &<Runtime as frame_system::Config>::RuntimeCall
 		MessageQueue(..) => (ON, ON),         // contains non-permissioned service calls
 		AssetRate(..) => (OFF, OFF),
 		Beefy(..) => (ON, ON), // For reporting equivocation proofs; security relevant
-		RcMigrator(..) => (ON, ON), // Required for the migration, only permissioned calls
 	}
 	// Exhaustive match. Compiler ensures that we did not miss any.
-}
-
-// Type safe mapping of RC hold reason to portable format.
-impl pallet_rc_migrator::types::IntoPortable for RuntimeHoldReason {
-	type Portable = pallet_rc_migrator::types::PortableHoldReason;
-
-	fn into_portable(self) -> Self::Portable {
-		use pallet_rc_migrator::types::PortableHoldReason;
-
-		match self {
-			RuntimeHoldReason::Preimage(inner) => PortableHoldReason::Preimage(inner),
-			RuntimeHoldReason::StateTrieMigration(inner) =>
-				PortableHoldReason::StateTrieMigration(inner),
-			RuntimeHoldReason::DelegatedStaking(inner) =>
-				PortableHoldReason::DelegatedStaking(inner),
-			RuntimeHoldReason::Staking(inner) => PortableHoldReason::Staking(inner),
-			RuntimeHoldReason::Session(inner) => PortableHoldReason::Session(inner),
-			RuntimeHoldReason::XcmPallet(inner) => PortableHoldReason::XcmPallet(inner),
-		}
-	}
-}
-
-impl pallet_rc_migrator::types::IntoPortable for RuntimeFreezeReason {
-	type Portable = pallet_rc_migrator::types::PortableFreezeReason;
-
-	fn into_portable(self) -> Self::Portable {
-		match self {
-			RuntimeFreezeReason::NominationPools(inner) =>
-				PortableFreezeReason::NominationPools(inner),
-		}
-	}
 }